refactor(go-consensus): split §23.2.4 deployment select logic into its own file [RUB-600]
Move descriptorFieldsValid + selectGoverningSurface + validDescriptorSubset +
governingDescriptor into simplicity_deployment_select.go so each file's aggregate
cyclomatic complexity stays under the Codacy 30 threshold (one file was 36). Pure
mechanical relocation of identical functions; no logic change.
